:symbols: Use relative path...
This importing types separate and then potentially lying about the type
is necessary because, near as I can find, the `"paths"` entries do not
respect relative paths in the `tsconfig.json` file and relative paths
are required to make this play nice with GitHub Pages.

Pull/Merge Request(s) that make this less gross would very much be
appreciated, because there likely will be other projects that'd not need
a bundle other than for reducing this sort of friction!